fix creation of linux shortcuts
