+++ openwebstart/src/test/java/com/openwebstart/os/mac/ScriptFactoryTest.java
@@ -1,12 +1,14 @@
 package com.openwebstart.os.mac;
 
-import com.openwebstart.os.ScriptFactory;
 import net.sourceforge.jnlp.JNLPFile;
 import net.sourceforge.jnlp.JNLPFileFactory;
 import org.junit.jupiter.api.Test;
 
 import java.net.URL;
 
+import static com.openwebstart.os.mac.ScriptFactory.createSimpleStartScriptForMac;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class ScriptFactoryTest {
@@ -15,13 +17,13 @@ class ScriptFactoryTest {
         final JNLPFileFactory f = new JNLPFileFactory();
         final URL jnlpUrl = this.getClass().getClassLoader()
                 .getResource("com/openwebstart/jnlp/with-http-location.jnlp");
-        assertTrue(jnlpUrl != null);
+        assertNotNull(jnlpUrl);
 
         final JNLPFile jnlp = f.create(jnlpUrl);
-        assertTrue(jnlp != null);
-        assertTrue("http://localhost/jnlp.jnlp".equals(jnlp.getSourceLocation().toString()));
+        assertNotNull(jnlp);
+        assertEquals(jnlp.getSourceLocation().toString(), "http://localhost/jnlp.jnlp");
 
-        final String script = ScriptFactory.createSimpleStartScriptForMac(jnlp);
+        final String script = createSimpleStartScriptForMac(jnlp);
         assertTrue(script.contains("open \"jnlp://localhost/jnlp.jnlp\""));
     }
 
@@ -30,13 +32,13 @@ class ScriptFactoryTest {
         final JNLPFileFactory f = new JNLPFileFactory();
         final URL jnlpUrl = this.getClass().getClassLoader()
                 .getResource("com/openwebstart/jnlp/with-https-location.jnlp");
-        assertTrue(jnlpUrl != null);
+        assertNotNull(jnlpUrl);
 
         final JNLPFile jnlp = f.create(jnlpUrl);
-        assertTrue(jnlp != null);
-        assertTrue("https://localhost/jnlp.jnlp".equals(jnlp.getSourceLocation().toString()));
+        assertNotNull(jnlp);
+        assertEquals(jnlp.getSourceLocation().toString(), "https://localhost/jnlp.jnlp");
 
-        final String script = ScriptFactory.createSimpleStartScriptForMac(jnlp);
+        final String script = createSimpleStartScriptForMac(jnlp);
         assertTrue(script.contains("open \"jnlps://localhost/jnlp.jnlp\""));
     }
 
@@ -45,13 +47,13 @@ class ScriptFactoryTest {
         final JNLPFileFactory f = new JNLPFileFactory();
         final URL jnlpUrl = this.getClass().getClassLoader()
                 .getResource("com/openwebstart/jnlp/with-file-location.jnlp");
-        assertTrue(jnlpUrl != null);
+        assertNotNull(jnlpUrl);
 
         final JNLPFile jnlp = f.create(jnlpUrl);
-        assertTrue(jnlp != null);
-        assertTrue("file:/path/jnlp.jnlp".equals(jnlp.getSourceLocation().toString()));
+        assertNotNull(jnlp);
+        assertEquals(jnlp.getSourceLocation().toString(), "file:/path/jnlp.jnlp");
 
-        final String script = ScriptFactory.createSimpleStartScriptForMac(jnlp);
+        final String script = createSimpleStartScriptForMac(jnlp);
         assertTrue(script.contains("open \"file:/path/jnlp.jnlp\""));
     }
 }
